+++ tests/Feature/UnitempDeletetPestTest.php
@@ -21,24 +21,24 @@ it('insert a record, retrive model and delete it', function () {
     expect($count)->toBe(1);
 
     $now = new Carbon;
-    $now = $now->format('Y-m-d H:i:s');
+    $now = $now->format('Y-m-d H:i:s.u');
 
     $rec1 = UniTemporalId::where('id', 1)->where('known_from', '<=', $now)->where('known_to', '>=', $now)->first();
 
     $rec1->text = 'Second Record';
     $rec1->save();
 
-    $maxDateTime = '9999-12-31 23:59:59';
+    $maxDateTime = '9999-12-31 23:59:59.000000';
 
     $recBeforeDel = UniTemporalId::withoutGlobalScope(UniTemporalScope::class)->where('id', 1)->latest('known_to')->first();
-    expect($recBeforeDel->known_to->format('Y-m-d H:i:s'))->toBe($maxDateTime);
+    expect($recBeforeDel->known_to->format('Y-m-d H:i:s.u'))->toBe($maxDateTime);
 
     sleep(1);
 
     $recBeforeDel->delete();
 
     $recAfterDel = UniTemporalId::withoutGlobalScope(UniTemporalScope::class)->where('id', 1)->latest('known_to')->first();
-    expect($recAfterDel->known_to->format('Y-m-d H:i:s'))->not->toBe($maxDateTime);
+    expect($recAfterDel->known_to->format('Y-m-d H:i:s.u'))->not->toBe($maxDateTime);
 
 });
 
@@ -52,23 +52,23 @@ it('insert a record, retrive model and delete it and check with latestVersion sc
     expect($count)->toBe(1);
 
     $now = new Carbon;
-    $now = $now->format('Y-m-d H:i:s');
+    $now = $now->format('Y-m-d H:i:s.u');
 
     $rec1 = UniTemporalId::where('id', 1)->where('known_from', '<=', $now)->where('known_to', '>=', $now)->first();
 
     $rec1->text = 'Second Record';
     $rec1->save();
 
-    $maxDateTime = '9999-12-31 23:59:59';
+    $maxDateTime = '9999-12-31 23:59:59.000000';
 
     $recBeforeDel = UniTemporalId::where('id', 1)->latestVersion()->first();
-    expect($recBeforeDel->known_to->format('Y-m-d H:i:s'))->toBe($maxDateTime);
+    expect($recBeforeDel->known_to->format('Y-m-d H:i:s.u'))->toBe($maxDateTime);
 
     sleep(1);
 
     $recBeforeDel->delete();
 
     $recAfterDel = UniTemporalId::where('id', 1)->latestVersion()->first();
-    expect($recAfterDel->known_to->format('Y-m-d H:i:s'))->not->toBe($maxDateTime);
+    expect($recAfterDel->known_to->format('Y-m-d H:i:s.u'))->not->toBe($maxDateTime);
 
 });